@@ -1,46 +1,55 @@
 {
   "name": "open-randomizer-tracker",
-  "version": "1.0.0",
+  "version": "0.1.0",
   "license": "BSD-4-Clause",
   "author": "Anthony Mattera",
   "private": true,
-  "main": "public/electron.js",
+  "main": "dist/electron.js",
   "dependencies": {
     "clone": "^2.1.2",
-    "electron-is-dev": "^1.1.0",
     "electron-store": "^3.2.0",
-    "node-sass": "^4.12.0",
-    "react": "^16.8.3",
-    "react-dom": "^16.8.3",
-    "react-scripts": "2.1.5",
-    "react-rnd": "^10.0.0",
     "require": "^2.4.20"
   },
   "devDependencies": {
-    "@rescripts/cli": "^0.0.10",
-    "@rescripts/rescript-env": "^0.0.5",
-    "concurrently": "^4.1.0",
+    "@babel/core": "^7.7.2",
+    "@babel/plugin-proposal-function-bind": "^7.2.0",
+    "@babel/preset-env": "^7.7.1",
+    "@babel/preset-react": "^7.7.0",
+    "@babel/preset-stage-0": "^7.0.0",
+    "babel-eslint": "^10.0.3",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-react-css-modules": "^5.2.6",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-react-jsx": "^6.24.1",
+    "copy-webpack-plugin": "^5.0.5",
+    "css-loader": "^3.2.0",
     "electron": "^4.0.6",
     "electron-builder": "^20.38.5",
-    "typescript": "^3.3.3333",
     "eslint": "^5.16.0",
     "eslint-config-google": "^0.13.0",
-    "eslint-plugin-react": "^7.13.0",
-    "typescript": "^3.3.3333",
-    "wait-on": "^3.2.0"
+    "eslint-loader": "^3.0.2",
+    "eslint-plugin-babel": "^5.3.0",
+    "eslint-plugin-react": "^7.16.0",
+    "file-loader": "^4.2.0",
+    "mini-css-extract-plugin": "^0.8.0",
+    "node-sass": "^4.13.0",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
+    "react-rnd": "^10.0.0",
+    "react-scripts": "2.1.5",
+    "sass-loader": "^8.0.0",
+    "wait-on": "^3.2.0",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0",
+    "webpack-glob-entry": "^2.1.1"
   },
   "scripts": {
-    "start": "rescripts start",
-    "build": "rescripts build",
-    "test": "rescripts test",
-    "electron-dev": "concurrently \"BROWSER=none yarn start\" \"wait-on http://localhost:3000 && electron .\"",
-    "postinstall": "electron-builder install-app-deps",
-    "preelectron-pack": "yarn build",
-    "electron-pack": "build -l"
+    "start": "webpack-dev-server --open",
+    "electron": "electron .",
+    "fix": "lint --fix",
+    "build": "webpack"
   },
-  "rescripts": [
-    "env"
-  ],
   "eslintConfig": {
     "extends": "react-app"
   },
