@@ -10,7 +10,7 @@ export default class ModuleLayout_LinkToThePast_Keysanity extends React.Componen
      * componentWillMount - Handles clearing the state and resetting it when the component will be mounted.
      * @return {void}
      */
-    componentWillMount() {
+    UNSAFE_componentWillMount() {
 
         const config = this.props.config;
         const state = {items:{}, dungeons:{}};
@@ -265,7 +265,7 @@ export default class ModuleLayout_LinkToThePast_Keysanity extends React.Componen
 
             // Assemble
             return <div key={key}>
-                <img className={classInactive} onClick={onClick} onContextMenu={onRightClick} src={window.location.origin + `/resources/modules/lttp/items/${icon}`}/>
+                <img className={classInactive} onClick={onClick} onContextMenu={onRightClick} src={`resources/modules/lttp/items/${icon}`}/>
                 {elementCounter}
             </div>;
 
@@ -321,7 +321,7 @@ export default class ModuleLayout_LinkToThePast_Keysanity extends React.Componen
                 const inactive = state.dungeons[key].active ? '' : 'inactive';
                 const icon = this.markerEnum[state.dungeons[key].marker] + '.png';
 
-                element = <img className={inactive} src={window.location.origin + '/resources/modules/lttp/icons/' + icon}/>
+                element = <img className={inactive} src={'resources/modules/lttp/icons/' + icon}/>
                 elementClass = 'action';
 
                 onLeftClick = () => this.onMarkerLeftClick(key, dungeon);
@@ -348,7 +348,7 @@ export default class ModuleLayout_LinkToThePast_Keysanity extends React.Componen
             if (dungeon.master !== false) {
 
                 const inactive = state.dungeons[key].master ? '' : 'inactive';
-                element = <img className={inactive} src={window.location.origin + '/resources/modules/lttp/icons/master.png'}/>;
+                element = <img className={inactive} src={'resources/modules/lttp/icons/master.png'}/>;
                 elementClass = 'action';
 
                 onLeftClick = () => this.onMasterLeftClick(key, dungeon);
@@ -385,7 +385,7 @@ export default class ModuleLayout_LinkToThePast_Keysanity extends React.Componen
                     classState = 'maxed';
                 }
 
-                element = <img className={classState} src={window.location.origin + '/resources/modules/lttp/icons/key.png'}/>;
+                element = <img className={classState} src={'resources/modules/lttp/icons/key.png'}/>;
                 elementText = <span className={classState}>{keys}/{dungeon.keys}</span>;
                 elementClass = 'action';
 
