@@ -1,6 +1,5 @@
 // May not need this in react land, but still want to be sure.
 require('events').EventEmitter.defaultMaxListeners = 99;
-const isDev = require('electron-is-dev');
 const {app, BrowserWindow} = require('electron');
 let mainWindow;
 
@@ -10,17 +9,17 @@ app.disableHardwareAcceleration();
 function createWindow () {
 
     if (process.platform === "win32") {
-        ico = 'app/resources/icons/icon.png'; 
+        ico = __dirname+'/resources/icons/icon.png'; 
     } else if (process.platform === "darwin") {
-        ico = 'app/resources/icons/icon.icns';
+        ico = __dirname+'/resources/icons/icon.icns';
     }else {
-        ico = 'app/resources/icons/icon.png';
+        ico = __dirname+'/resources/icons/icon.png';
     } 
     
     const windowConfig = {
         icon:ico,
-        width:320, 
-        height:200, 
+        width:800, 
+        height:600, 
         x:0, 
         y:0, 
         minWidth:50, 
@@ -36,8 +35,7 @@ function createWindow () {
     mainWindow.setMenu(null);
 
     // Our Controller File
-    mainWindow.loadURL(isDev ? 'http://localhost:3000' : `file://${path.join(__dirname, '../build/index.html')}`);
-    //mainWindow.loadFile('index.html');
+    mainWindow.loadFile(__dirname+'/index.html');
     
 }
 
